backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
asyncpg==0.29.0
SQLAlchemy==2.0.25
GeoAlchemy2==0.14.3
alembic==1.13.1
pydantic==2.6.1
pydantic-settings==2.1.0
redis==5.0.1
websockets==12.0
psycopg2-binary
python-jose
passlib[bcrypt]
python-multipart
pytest
httpx
pytest-asyncio
